requirements.txt
streamlit
pandas
numpy
nltk
sentence-transformers
faiss-cpu
groq
newsapi-python
newspaper3k
lxml[html_clean]
tqdm
snscrape